+++ backend/Drinker/src/main/java/com/Drinker/repository/MatchRepo.java
@@ -4,6 +4,6 @@ import com.Drinker.model.Match;
 import com.Drinker.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepo extends JpaRepository<Match, Long> {
+public interface MatchRepo extends JpaRepository<Match, Integer> {
     Match findByUser2AndUser1(User user2, User user1);
 }
