+++ backend/Drinker/src/main/java/com/Drinker/repository/PlaceRepo.java
@@ -3,5 +3,5 @@ package com.Drinker.repository;
 import com.Drinker.model.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaceRepo extends JpaRepository<Place, Long> {
+public interface PlaceRepo extends JpaRepository<Place, Integer> {
 }
